Cli support (#2)
* Initial commit

* Add src folder and basic __ files

* Add pre-commit and support with yapf mirror

* Add vim to .gitignore

* Add pylint to pre-commit

* Add setup script to install required tools

* Add setup.py script

* Update basic pathing

* Add testing with tox and pytest

* Fix spacing issue with testing example

* Add code coverage checks using pytest-cov

* Start adding a checklist users of the template should follow

* Attempt to add CI test with github actions and tox

* Add style and fix versioning

* Fix missing

* Add test badge to the Readme

* Add Checklist item for running installation (#4)

* Simplify tox dependencies (#5)

This does not address the issues in an expected way. The issue recommends that we update the documentation about how to add dependencies to the tox.ini file. However, I felt that this still requires the user to maintain multiple dependency lists in multiple locations, thus not simplifying the issue. I have utilised a tox.ini feature to move the dependency list into the setup.py file.

I also believe that the setup.py repository should be understood by general python programmers and thus we do not need to explain this change in the checklist.

* Update template application (#6)

* Update example template to a hello world

* Add test run example

* Click support with basic cli

Co-authored-by: kyle <ljlabs@users.noreply.github.com>